Remove types.ts and update imports for type definitions
Deleted src/types.ts and refactored code in src/index.ts and src/plakar.ts to inline or locally define types previously imported from types.ts. This simplifies the codebase by reducing indirection and consolidating type definitions.
