+++ src/plakar.ts
@@ -1,43 +1,403 @@
 import { spawn, SpawnOptions } from "child_process";
-import {
-  CommonOptions,
-  BackupOptions,
-  ListOptions,
-  RestoreOptions,
-  CreateOptions,
-  SyncOptions,
-  CheckOptions,
-  CloneOptions,
-  MountOptions,
-  UIOptions,
-  ServerOptions,
-  RemoveOptions,
-  DiffOptions,
-  LocateOptions,
-  CatOptions,
-  ArchiveOptions,
-  ConfigRepositoryOptions,
-  MaintenanceOptions,
-  Snapshot,
-  AgentStatus,
-  VersionInfo,
-  CommandResult,
-} from "./types";
+// import {
+//   CommonOptions,
+//   BackupOptions,
+//   ListOptions,
+//   RestoreOptions,
+//   CreateOptions,
+//   SyncOptions,
+//   CheckOptions,
+//   CloneOptions,
+//   MountOptions,
+//   UIOptions,
+//   ServerOptions,
+//   RemoveOptions,
+//   DiffOptions,
+//   LocateOptions,
+//   CatOptions,
+//   ArchiveOptions,
+//   ConfigRepositoryOptions,
+//   MaintenanceOptions,
+//   Snapshot,
+//   AgentStatus,
+//   VersionInfo,
+//   CommandResult,
+// } from "./types";
 import { PlakarInstaller, InstallOptions, CheckResult, InstallResult } from "./installer";
+
+/**
+ * Plakar CLI wrapper for Node.js
+ * Provides a TypeScript interface to interact with the Plakar backup tool.
+ * Supports all major commands and options.
+ *
+ * @module plakar
+ */
+namespace Plakar{
+  /**
+   * Common options available for most Plakar commands
+   */
+  export interface CommonOptions {
+    /** Use the configuration at path */
+    config?: string;
+    /** Limit the number of parallelism */
+    cpu?: number;
+    /** Change the hostname used for backups */
+    hostname?: string;
+    /** Use the passphrase from the key file */
+    keyfile?: string;
+    /** Run without attempting to connect to the agent */
+    noAgent?: boolean;
+    /** Disable all output except for errors */
+    quiet?: boolean;
+    /** Display trace logs */
+    trace?: string;
+    /** Change the username used for backups */
+    username?: string;
+  }
+
+  /**
+   * Repository specification
+   */
+  export interface Repository {
+    /** Repository path, URI, or label */
+    location: string;
+  }
+
+  /**
+   * Backup options
+   */
+  export interface BackupOptions extends CommonOptions {
+    /** Set the maximum number of parallel tasks */
+    concurrency?: number;
+    /** Specify individual glob exclusion patterns */
+    exclude?: string[];
+    /** Specify a file containing glob exclusion patterns */
+    excludes?: string;
+    /** Perform a full check on the backup after success */
+    check?: boolean;
+    /** Specify a tag to assign to the snapshot */
+    tag?: string;
+  }
+
+  /**
+   * List options for filtering snapshots
+   */
+  export interface ListOptions extends CommonOptions {
+    /** Display the full UUID for each snapshot */
+    uuid?: boolean;
+    /** Filter by snapshot name */
+    name?: string;
+    /** Filter by category */
+    category?: string;
+    /** Filter by environment */
+    environment?: string;
+    /** Filter by perimeter */
+    perimeter?: string;
+    /** Filter by job */
+    job?: string;
+    /** Filter by tag */
+    tag?: string;
+    /** Only show latest snapshot matching filters */
+    latest?: boolean;
+    /** Only show snapshots older than the specified date */
+    before?: string;
+    /** Only show snapshots created since the specified date */
+    since?: string;
+    /** List directory contents recursively */
+    recursive?: boolean;
+  }
+
+  /**
+   * Restore options
+   */
+  export interface RestoreOptions extends CommonOptions {
+    /** Filter by snapshot name */
+    name?: string;
+    /** Filter by category */
+    category?: string;
+    /** Filter by environment */
+    environment?: string;
+    /** Filter by perimeter */
+    perimeter?: string;
+    /** Filter by job */
+    job?: string;
+    /** Filter by tag */
+    tag?: string;
+    /** Only restore from latest snapshot matching filters */
+    latest?: boolean;
+    /** Only restore from snapshots older than the specified date */
+    before?: string;
+    /** Only restore from snapshots created since the specified date */
+    since?: string;
+    /** Set the maximum number of parallel tasks */
+    concurrency?: number;
+    /** Strip the original path from each restored file */
+    rebase?: boolean;
+    /** Specify the base directory to which files will be restored */
+    to?: string;
+  }
+
+  /**
+   * Repository creation options
+   */
+  export interface CreateOptions extends CommonOptions {
+    /** Create unencrypted repository */
+    noEncryption?: boolean;
+  }
+
+  /**
+   * Sync options
+   */
+  export interface SyncOptions extends CommonOptions {
+    /** Synchronization direction */
+    direction?: "to" | "from" | "with";
+    /** Target repository for synchronization */
+    target: string;
+  }
+
+  /**
+   * Check options
+   */
+  export interface CheckOptions extends CommonOptions {
+    /** Perform deep integrity check */
+    deep?: boolean;
+  }
+
+  /**
+   * Clone options
+   */
+  export interface CloneOptions extends CommonOptions {
+    /** Target repository for cloning */
+    target: string;
+  }
+
+  /**
+   * Mount options
+   */
+  export interface MountOptions extends CommonOptions {
+    /** Mount point directory */
+    mountPoint: string;
+    /** Mount in read-only mode */
+    readonly?: boolean;
+  }
+
+  /**
+   * UI options
+   */
+  export interface UIOptions extends CommonOptions {
+    /** Port to bind the web interface */
+    port?: number;
+    /** Host to bind the web interface */
+    host?: string;
+  }
+
+  /**
+   * Server options
+   */
+  export interface ServerOptions extends CommonOptions {
+    /** Port to bind the server */
+    port?: number;
+    /** Host to bind the server */
+    host?: string;
+  }
+
+  /**
+   * Remove options
+   */
+  export interface RemoveOptions extends CommonOptions {
+    /** Remove snapshots older than the specified date */
+    before?: string;
+    /** Remove snapshots created since the specified date */
+    since?: string;
+    /** Filter by snapshot name */
+    name?: string;
+    /** Filter by category */
+    category?: string;
+    /** Filter by environment */
+    environment?: string;
+    /** Filter by perimeter */
+    perimeter?: string;
+    /** Filter by job */
+    job?: string;
+    /** Filter by tag */
+    tag?: string;
+    /** Force removal without confirmation */
+    force?: boolean;
+  }
+
+  /**
+   * Snapshot information
+   */
+  export interface Snapshot {
+    /** Snapshot ID */
+    id: string;
+    /** Full UUID */
+    uuid?: string;
+    /** Creation timestamp */
+    timestamp: string;
+    /** Snapshot size */
+    size: string;
+    /** Duration of backup */
+    duration: string;
+    /** Backed up path */
+    path: string;
+    /** Tags associated with the snapshot */
+    tags?: string[];
+  }
+
+  /**
+   * File information in a snapshot
+   */
+  export interface FileInfo {
+    /** File path */
+    path: string;
+    /** File size */
+    size: number;
+    /** File permissions */
+    mode: string;
+    /** File modification time */
+    mtime: string;
+    /** File type */
+    type: "file" | "directory" | "symlink";
+    /** Link target for symlinks */
+    target?: string;
+  }
+
+  /**
+   * Repository information
+   */
+  export interface RepositoryInfo {
+    /** Repository location */
+    location: string;
+    /** Repository type */
+    type: string;
+    /** Whether repository is encrypted */
+    encrypted: boolean;
+    /** Number of snapshots */
+    snapshots: number;
+    /** Total size */
+    totalSize: string;
+  }
+
+  /**
+   * Agent status
+   */
+  export interface AgentStatus {
+    /** Whether agent is running */
+    running: boolean;
+    /** Agent process ID */
+    pid?: number;
+  }
+
+  /**
+   * Version information
+   */
+  export interface VersionInfo {
+    /** Plakar version */
+    version: string;
+    /** Available importers */
+    importers: string[];
+    /** Available exporters */
+    exporters: string[];
+    /** Available klosets */
+    klosets: string[];
+  }
+
+  /**
+   * Command execution result
+   */
+  export interface CommandResult {
+    /** Exit code */
+    exitCode: number;
+    /** Standard output */
+    stdout: string;
+    /** Standard error */
+    stderr: string;
+    /** Whether command succeeded */
+    success: boolean;
+  }
+
+  /**
+   * Diff options
+   */
+  export interface DiffOptions extends CommonOptions {
+    /** Show context lines */
+    context?: number;
+    /** Show unified diff */
+    unified?: boolean;
+  }
+
+  /**
+   * Locate options
+   */
+  export interface LocateOptions extends CommonOptions {
+    /** Pattern to search for */
+    pattern: string;
+    /** Case-sensitive search */
+    caseSensitive?: boolean;
+  }
+
+  /**
+   * Cat options
+   */
+  export interface CatOptions extends CommonOptions {
+    /** Show line numbers */
+    number?: boolean;
+  }
+
+  /**
+   * Archive options
+   */
+  export interface ArchiveOptions extends CommonOptions {
+    /** Output format */
+    format?: "tar" | "zip";
+    /** Output file path */
+    output: string;
+    /** Compression level */
+    compression?: number;
+  }
+
+  /**
+   * Config repository options
+   */
+  export interface ConfigRepositoryOptions {
+    /** Repository name */
+    name: string;
+    /** Repository location */
+    location?: string;
+    /** Repository passphrase */
+    passphrase?: string;
+    /** Access key for S3 repositories */
+    accessKey?: string;
+    /** Secret access key for S3 repositories */
+    secretAccessKey?: string;
+  }
+
+  /**
+   * Maintenance options
+   */
+  export interface MaintenanceOptions extends CommonOptions {
+    /** Perform dry run */
+    dryRun?: boolean;
+    /** Force maintenance */
+    force?: boolean;
+  }
+
+}
 
 /**
  * Main Plakar wrapper class
  */
-export class Plakar {
+class Plakar {
   private binaryPath: string;
   private defaultRepository?: string;
-  private defaultOptions: CommonOptions;
+  private defaultOptions: Plakar.CommonOptions;
 
   constructor(
     options: {
       binaryPath?: string;
       repository?: string;
-      defaultOptions?: CommonOptions;
+      defaultOptions?: Plakar.CommonOptions;
       autoInstall?: boolean;
     } = {},
   ) {
@@ -56,7 +416,7 @@ export class Plakar {
   /**
    * Execute a plakar command
    */
-  private async executeCommand(args: string[], options: CommonOptions = {}): Promise<CommandResult> {
+  private async executeCommand(args: string[], options: Plakar.CommonOptions = {}): Promise<Plakar.CommandResult> {
     const mergedOptions = { ...this.defaultOptions, ...options };
     const commandArgs = this.buildCommandArgs(args, mergedOptions);
 
@@ -90,7 +450,7 @@ export class Plakar {
       });
 
       child.on("close", (code) => {
-        const result: CommandResult = {
+        const result: Plakar.CommandResult = {
           exitCode: code || 0,
           stdout: stdout.trim(),
           stderr: stderr.trim(),
@@ -115,7 +475,7 @@ export class Plakar {
   /**
    * Build command arguments from options
    */
-  private buildCommandArgs(baseArgs: string[], options: CommonOptions): string[] {
+  private buildCommandArgs(baseArgs: string[], options: Plakar.CommonOptions): string[] {
     const args = [...baseArgs];
 
     if (options.config) {
@@ -159,9 +519,9 @@ export class Plakar {
   /**
    * Parse snapshot list output
    */
-  private parseSnapshotList(output: string): Snapshot[] {
+  private parseSnapshotList(output: string): Plakar.Snapshot[] {
     const lines = output.split("\n").filter((line) => line.trim());
-    const snapshots: Snapshot[] = [];
+    const snapshots: Plakar.Snapshot[] = [];
 
     for (const line of lines) {
       // Parse format: "2025-02-19T21:38:16Z   9abc3294    3.1 MB      0s   /private/etc"
@@ -183,12 +543,12 @@ export class Plakar {
   /**
    * Parse version output
    */
-  private parseVersion(output: string): VersionInfo {
+  private parseVersion(output: string): Plakar.VersionInfo {
     const lines = output.split("\n");
     const versionLine = lines[0] || "";
     const version = versionLine.replace(/^plakar\//, "");
 
-    const info: VersionInfo = {
+    const info: Plakar.VersionInfo = {
       version,
       importers: [],
       exporters: [],
@@ -211,7 +571,7 @@ export class Plakar {
   /**
    * Start the plakar agent
    */
-  async startAgent(options: CommonOptions = {}): Promise<AgentStatus> {
+  async startAgent(options: Plakar.CommonOptions = {}): Promise<Plakar.AgentStatus> {
     const result = await this.executeCommand(["agent"], options);
 
     // Parse agent output to extract PID
@@ -227,14 +587,14 @@ export class Plakar {
   /**
    * Stop the plakar agent
    */
-  async stopAgent(options: CommonOptions = {}): Promise<void> {
+  async stopAgent(options: Plakar.CommonOptions = {}): Promise<void> {
     await this.executeCommand(["agent", "stop"], options);
   }
 
   /**
    * Get agent status
    */
-  async getAgentStatus(options: CommonOptions = {}): Promise<AgentStatus> {
+  async getAgentStatus(options: Plakar.CommonOptions = {}): Promise<Plakar.AgentStatus> {
     try {
       const result = await this.executeCommand(["agent", "status"], options);
       const pidMatch = result.stdout.match(/pid=(\d+)/);
@@ -251,7 +611,7 @@ export class Plakar {
   /**
    * Create a new repository
    */
-  async createRepository(repository?: string, options: CreateOptions = {}): Promise<void> {
+  async createRepository(repository?: string, options: Plakar.CreateOptions = {}): Promise<void> {
     const args = ["create"];
     if (options.noEncryption) {
       args.push("-no-encryption");
@@ -263,7 +623,7 @@ export class Plakar {
   /**
    * Create a backup snapshot
    */
-  async backup(path?: string, repository?: string, options: BackupOptions = {}): Promise<string> {
+  async backup(path?: string, repository?: string, options: Plakar.BackupOptions = {}): Promise<string> {
     const args = ["backup"];
 
     if (options.concurrency) {
@@ -297,7 +657,7 @@ export class Plakar {
   /**
    * List snapshots or snapshot contents
    */
-  async list(snapshotPath?: string, repository?: string, options: ListOptions = {}): Promise<Snapshot[]> {
+  async list(snapshotPath?: string, repository?: string, options: Plakar.ListOptions = {}): Promise<Plakar.Snapshot[]> {
     const args = ["ls"];
 
     if (options.uuid) args.push("-uuid");
@@ -321,7 +681,7 @@ export class Plakar {
   /**
    * Restore files from a snapshot
    */
-  async restore(snapshotPath?: string, repository?: string, options: RestoreOptions = {}): Promise<void> {
+  async restore(snapshotPath?: string, repository?: string, options: Plakar.RestoreOptions = {}): Promise<void> {
     const args = ["restore"];
 
     if (options.name) args.push("-name", options.name);
@@ -344,7 +704,7 @@ export class Plakar {
   /**
    * Check repository integrity
    */
-  async check(snapshotId?: string, repository?: string, options: CheckOptions = {}): Promise<void> {
+  async check(snapshotId?: string, repository?: string, options: Plakar.CheckOptions = {}): Promise<void> {
     const args = ["check"];
     if (snapshotId) args.push(snapshotId);
 
@@ -354,7 +714,7 @@ export class Plakar {
   /**
    * Synchronize repositories
    */
-  async sync(target: string, repository?: string, options: Omit<SyncOptions, "target"> = {}): Promise<void> {
+  async sync(target: string, repository?: string, options: Omit<Plakar.SyncOptions, "target"> = {}): Promise<void> {
     const direction = options.direction || "to";
     const args = ["sync", direction, target];
 
@@ -364,7 +724,7 @@ export class Plakar {
   /**
    * Clone a repository
    */
-  async clone(target: string, repository?: string, options: Omit<CloneOptions, "target"> = {}): Promise<void> {
+  async clone(target: string, repository?: string, options: Omit<Plakar.CloneOptions, "target"> = {}): Promise<void> {
     const args = ["clone", target];
 
     await this.executeCommand(this.addRepositoryArgs(args, repository), options);
@@ -373,7 +733,7 @@ export class Plakar {
   /**
    * Display file contents from a snapshot
    */
-  async cat(snapshotPath: string, repository?: string, options: CatOptions = {}): Promise<string> {
+  async cat(snapshotPath: string, repository?: string, options: Plakar.CatOptions = {}): Promise<string> {
     const args = ["cat"];
     if (options.number) args.push("-number");
     args.push(snapshotPath);
@@ -386,7 +746,7 @@ export class Plakar {
   /**
    * Show differences between files
    */
-  async diff(snapshot1: string, snapshot2: string, repository?: string, options: DiffOptions = {}): Promise<string> {
+  async diff(snapshot1: string, snapshot2: string, repository?: string, options: Plakar.DiffOptions = {}): Promise<string> {
     const args = ["diff"];
     if (options.context) args.push("-context", options.context.toString());
     if (options.unified) args.push("-unified");
@@ -400,7 +760,7 @@ export class Plakar {
   /**
    * Find filenames in snapshots
    */
-  async locate(pattern: string, snapshotId?: string, repository?: string, options: Omit<LocateOptions, "pattern"> = {}): Promise<string[]> {
+  async locate(pattern: string, snapshotId?: string, repository?: string, options: Omit<Plakar.LocateOptions, "pattern"> = {}): Promise<string[]> {
     const args = ["locate"];
     if (options.caseSensitive) args.push("-case-sensitive");
     args.push(pattern);
@@ -414,7 +774,7 @@ export class Plakar {
   /**
    * Remove snapshots
    */
-  async remove(snapshotIds?: string[], repository?: string, options: RemoveOptions = {}): Promise<void> {
+  async remove(snapshotIds?: string[], repository?: string, options: Plakar.RemoveOptions = {}): Promise<void> {
     const args = ["rm"];
 
     if (options.before) args.push("-before", options.before);
@@ -434,7 +794,7 @@ export class Plakar {
   /**
    * Mount snapshots as filesystem
    */
-  async mount(mountPoint: string, snapshotId?: string, repository?: string, options: Omit<MountOptions, "mountPoint"> = {}): Promise<void> {
+  async mount(mountPoint: string, snapshotId?: string, repository?: string, options: Omit<Plakar.MountOptions, "mountPoint"> = {}): Promise<void> {
     const args = ["mount"];
     if (options.readonly) args.push("-readonly");
     args.push(mountPoint);
@@ -446,7 +806,7 @@ export class Plakar {
   /**
    * Start web UI
    */
-  async startUI(repository?: string, options: UIOptions = {}): Promise<void> {
+  async startUI(repository?: string, options: Plakar.UIOptions = {}): Promise<void> {
     const args = ["ui"];
     if (options.port) args.push("-port", options.port.toString());
     if (options.host) args.push("-host", options.host);
@@ -457,7 +817,7 @@ export class Plakar {
   /**
    * Start server
    */
-  async startServer(repository?: string, options: ServerOptions = {}): Promise<void> {
+  async startServer(repository?: string, options: Plakar.ServerOptions = {}): Promise<void> {
     const args = ["server"];
     if (options.port) args.push("-port", options.port.toString());
     if (options.host) args.push("-host", options.host);
@@ -468,7 +828,7 @@ export class Plakar {
   /**
    * Create archive from snapshot
    */
-  async archive(snapshotPath: string, output: string, repository?: string, options: Omit<ArchiveOptions, "output"> = {}): Promise<void> {
+  async archive(snapshotPath: string, output: string, repository?: string, options: Omit<Plakar.ArchiveOptions, "output"> = {}): Promise<void> {
     const args = ["archive"];
     if (options.format) args.push("-format", options.format);
     if (options.compression) args.push("-compression", options.compression.toString());
@@ -480,7 +840,7 @@ export class Plakar {
   /**
    * Perform repository maintenance
    */
-  async maintenance(repository?: string, options: MaintenanceOptions = {}): Promise<void> {
+  async maintenance(repository?: string, options: Plakar.MaintenanceOptions = {}): Promise<void> {
     const args = ["maintenance"];
     if (options.dryRun) args.push("-dry-run");
     if (options.force) args.push("-force");
@@ -491,7 +851,7 @@ export class Plakar {
   /**
    * Get repository information
    */
-  async info(snapshotId?: string, repository?: string, options: CommonOptions = {}): Promise<string> {
+  async info(snapshotId?: string, repository?: string, options: Plakar.CommonOptions = {}): Promise<string> {
     const args = ["info"];
     if (snapshotId) args.push(snapshotId);
 
@@ -503,7 +863,7 @@ export class Plakar {
   /**
    * Get version information
    */
-  async version(options: CommonOptions = {}): Promise<VersionInfo> {
+  async version(options: Plakar.CommonOptions = {}): Promise<Plakar.VersionInfo> {
     const result = await this.executeCommand(["version"], options);
     return this.parseVersion(result.stdout);
   }
@@ -511,7 +871,7 @@ export class Plakar {
   /**
    * Configuration management
    */
-  async configCreateRepository(name: string, options: ConfigRepositoryOptions): Promise<void> {
+  async configCreateRepository(name: string, options: Plakar.ConfigRepositoryOptions): Promise<void> {
     await this.executeCommand(["config", "repository", "create", name]);
 
     if (options.location) {
@@ -603,7 +963,7 @@ export class Plakar {
   /**
    * Raw command execution for advanced usage
    */
-  async raw(args: string[], options: CommonOptions = {}): Promise<CommandResult> {
+  async raw(args: string[], options: Plakar.CommonOptions = {}): Promise<Plakar.CommandResult> {
     // Check if Plakar is available before executing command
     const checkResult = await this.checkInstallation();
     if (!checkResult.installed) {
@@ -613,3 +973,5 @@ export class Plakar {
     return this.executeCommand(args, options);
   }
 }
+
+export { Plakar };
